% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drat_builder.R
\name{install_script}
\alias{install_script}
\title{Install drat.builder script}
\usage{
install_script(path)
}
\arguments{
\item{path}{Directory to install the script}
}
\description{
Installs the "drat.builder" script in the directory \code{path}
}

